split_kernels header + other changes
