+++ include/groute/worklist/work_target.cu.h
@@ -30,7 +30,7 @@
 #ifndef __GROUTE_WORK_TARGET_H
 #define __GROUTE_WORK_TARGET_H
 
-#include <groute/worklist/work_queue.cu.h>
+#include <groute/device/queue.cu.h>
 #include <groute/worklist/distributed_worklist.cu.h>
 
 namespace groute {
